Changed it all to longs

+++ LargestPrime/src/inialObservation/FindPrimes.java
@@ -1,18 +1,16 @@
 package inialObservation;
 
-import javax.swing.JOptionPane;
-
 public class FindPrimes {
 	
-	public static void simplePrimeUpto(int upto) {
+	public static void simplePrimeUpto(long upto) {
 		
-		//first loop, go through ints
-		for(int i = 1; i <= upto; i += 2) {
+		//first loop, go through longs
+		for(long i = 1; i <= upto; i += 2) {
 			
 			boolean isPrime = true;
 			
 			//second loop, see if divisible by numbers less than it
-			for(int j = 2; j < i / 2; j++) {
+			for(long j = 2; j < i / 2; j++) {
 				
 				if(i % j == 0) {
 					isPrime = false;
@@ -27,8 +25,8 @@ public class FindPrimes {
 	}
 	
 	public static void main(String[] args) {
-		//int upto = Integer.parseInt(JOptionPane.showInputDialog("Enter a number to go upto"));
-		int upto = Integer.MAX_VALUE;
+		//long upto = Long.parseInt(JOptionPane.showInputDialog("Enter a number to go upto"));
+		long upto = Long.MAX_VALUE;
 		simplePrimeUpto(upto);
 	}
 }
